+++ gui/next/src/lib/ui/Header.svelte
@@ -26,40 +26,98 @@ const preloadTables = async () => {
 <!-- ================================================================== -->
 <style>
 
+/* layout */
 header {
-  padding-block: 1rem;
-  height: 82px;
+  max-width: 100vw;
+  padding-block: var(--space-navigation);
   position: sticky;
   top: 0;
   z-index: 100;
 
   border-bottom: 1px solid var(--color-frame);
-  background-color: rgba(var(--color-rgb-page), .5);
-  backdrop-filter: blur(17px);
-  -webkit-backdrop-filter: blur(17px);
-
-  transition: filter .2s ease-in-out;
+  background-color: var(--color-page);
 }
 
 .wrapper {
-  margin-inline: auto;
-  padding-inline: 2rem;
+  min-width: 0;
+  width: 100%;
+  padding-inline: var(--space-page);
   display: flex;
   justify-content: space-between;
   align-items: center;
-  gap: 2em;
+  flex-wrap: wrap;
+  gap: var(--space-navigation) var(--space-page);
 }
 
-  @media (max-width: 570px) {
-    .wrapper {
-      flex-direction: column;
-    }
+/* logo */
+.logo {
+  min-width: 0;
+  min-height: 2.5625rem;
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.logo .label {
+  position: absolute;
+  left: -100vw;
+}
+
+.logo .sign {
+  width: 100%;
+  min-width: 2rem;
+  max-width: 3.125rem;
+
+  transition: scale .2s var(--easing-rapid);
+}
+
+  .logo .sign:hover,
+  .logo:has(.logotype:hover) .sign {
+    scale: 1.1;
   }
 
-h1 {
-  line-height: .5em;
+.logo h1 {
+  min-width: 2rem;
+  display: flex;
+  flex-direction: column;
 }
 
+.logo .logotype {
+  width: 100%;
+  max-width: 120px;
+
+  fill: var(--color-text);
+
+  transition: fill .2s var(--easing-rapid);
+}
+
+  .logo .logotype:hover,
+  .logo:has(.sign:hover) .logotype {
+    fill: color-mix(in srgb, var(--color-text), var(--color-text-secondary) 40%);
+  }
+
+.logo .instance {
+  max-width: 260px;
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  font-size: .8rem;
+  color: var(--color-text-secondary);
+
+  transition: font-size .2s ease-in-out;
+}
+
+.logo .instance.offline {
+  font-size: 0;
+  color: transparent;
+}
+
+.logo .instance:hover {
+  color: var(--color-interaction-hover);
+}
+
+/* navigation */
 ul {
   display: flex;
   gap: 1rem;
@@ -86,6 +144,12 @@ li a {
   transition-timing-function: linear, cubic-bezier(0.175, 0.885, 0.32, 1.275);
 }
 
+  @media (max-width: 525px) {
+    li a {
+      padding: .5rem;
+    }
+  }
+
   li a:hover {
     border-radius: 1.2rem;
     background-color: var(--color-middleground);
@@ -98,12 +162,14 @@ li a {
   }
 
 /* tooltip appearing on hover */
-.label {
+nav .label {
   margin-block-start: .4rem;
   padding: .2rem .5rem;
   position: absolute;
   top: 105%;
   right: 0;
+  left: auto;
+  bottom: auto;
   opacity: 0;
 
   border-radius: .2rem;
@@ -116,7 +182,7 @@ li a {
   transition: opacity .1s ease-in-out;
 }
 
-.label:before {
+nav .label:before {
   width: 10px;
   height: 6px;
   position: absolute;
@@ -129,7 +195,7 @@ li a {
   content: '';
 }
 
-a:hover .label {
+nav a:hover .label {
   opacity: 1;
 }
 
@@ -141,13 +207,27 @@ a:hover .label {
 <header>
   <div class="wrapper">
 
-    <a href="/">
+    <div class="logo">
+      <a href="/">
+        <svg class="sign" xmlns="http://www.w3.org/2000/svg" xml:space="preserve" x="0" y="0" style="enable-background:new 0 0 28.4 24.5" version="1.1" viewBox="0 0 28.4 24.5"><style>.st0,.st1{fill-rule:evenodd;clip-rule:evenodd;fill:#56b146}.st1{fill:#c3233d}.st2,.st3,.st4,.st5,.st6,.st7,.st8,.st9{fill-rule:evenodd;clip-rule:evenodd;fill:#d5246a}.st3,.st4,.st5,.st6,.st7,.st8,.st9{fill:#e77a2b}.st4,.st5,.st6,.st7,.st8,.st9{fill:#fac922}.st5,.st6,.st7,.st8,.st9{fill:#487fb5}.st6,.st7,.st8,.st9{fill:#23477b}.st7,.st8,.st9{fill:#419845}.st8,.st9{fill:#014c3d}.st9{fill:#008b47}</style><path d="M4.7 0h2.4L3.3 3.9 4.7 0z" class="st0"/><path d="M2 10.4 0 6.8h2v3.6z" class="st1"/><path d="M2 6.7h4l-4 3.6V6.7z" class="st2"/><path d="M5.8 13.6 6 6.7l-4 3.6 3.8 3.3z" class="st3"/><path d="m10.6 10.2-4.8 3.4.2-6.9 4.6 3.5z" class="st4"/><path d="M13.6 5.7 6 6.7l4.7 3.4 2.9-4.4z" class="st0"/><path d="M11.9 0 8.8 3.8l4.9 1.8L11.9 0z" class="st1"/><path d="M19.1 0h-7.3l1.8 5.7L19.1 0z" class="st5"/><path d="m6 6.7 7.6-1.1-4.8-1.8L6 6.7z" class="st2"/><path d="M20.5 5.7h-6.9L19.1 0l1.4 5.7z" class="st6"/><path d="m17 12-6.4-1.8 3-4.5L17 12z" class="st7"/><path d="M20.5 5.7 17 12l-3.4-6.3h6.9z" class="st8"/><path d="M22.9 10.2 17 12l3.4-6.3 2.5 4.5z" class="st7"/><path d="M21.8 13.9 17 12l5.8-1.8-1 3.7z" class="st0"/><path d="m19.1 0 5.3 4.4-3.9 1.3L19.1 0z" class="st5"/><path d="m20.5 5.7 3.9-1.3-1.5 5.8-2.4-4.5z" class="st9"/><path d="m22.9 10.2 5.4 1-6.5 2.6 1.1-3.6z" class="st8"/><path d="m28.3 11.2.1 5.2-6.5-2.6 6.4-2.6z" class="st7"/><path d="m21.8 13.9 6.5 2.5-4.9.8-1.6-3.3z" class="st6"/><path d="m28.4 16.4-2.5 5.7-2.4-4.8 4.9-.9z" class="st0"/><defs><filter id="Adobe_OpacityMaskFilter" width="10.8" height="7.2" x="15.1" y="17.2" filterUnits="userSpaceOnUse"><feColorMatrix values="1 0 0 0 0 0 1 0 0 0 0 0 1 0 0 0 0 0 1 0"/></filter></defs><mask id="a_00000065067817451968692310000014300811633783240372_" width="10.8" height="7.2" x="15.1" y="17.2" maskUnits="userSpaceOnUse"><path d="M15.1 17.2h10.8v7.2H15.1v-7.2z" style="fill-rule:evenodd;clip-rule:evenodd;fill:#fff;filter:url(#Adobe_OpacityMaskFilter)"/></mask><g style="mask:url(#a_00000065067817451968692310000014300811633783240372_)"><path d="m23.4 17.2 2.4 4.8-3.6-2 1.2-2.8z" class="st5"/><path d="m25.8 22.1-3.6 2.4V20l3.6 2.1z" class="st6"/><path d="M22.2 20v4.5l-1.8-2.9 1.8-1.6z" class="st5"/><path d="M22.2 24.5h-4.8l3-2.9 1.8 2.9z" class="st6"/><path d="m20.4 21.6-3 2.9.5-2.9h2.5z" class="st1"/><path d="m17.9 21.6-.5 2.9-2.3-2.9h2.8z" class="st2"/></g><path d="M17.9 21.6H15l1.7-2 1.2 2z" class="st1"/><path d="M15.1 21.6v-3.8l1.7 1.8-1.7 2z" class="st3"/><path d="m24.4 4.4 3.9 6.8-5.4-1 1.5-5.8z" class="st0"/><path d="m17.9 17.9-.4-1.9 2.9 1.3-2.5.6z" class="st7"/><path d="m15.1 17.8 2.9.1-.4-1.9-2.5 1.8z" class="st0"/><path d="m16.7 19.6 1.2-1.7-2.9-.1 1.7 1.8z" class="st4"/><path d="m6.7 2.5 2.1 1.3-.1-2.6-2 1.3z" class="st7"/><path d="m6.7 2.5 2-1.3L7.1 0l-.4 2.5z" class="st0"/><path d="M3.3 3.9 7.1 0l-.4 2.5-3.4 1.4z" class="st9"/><path d="M8.8 3.8 6.7 2.5l-1 1.4 3.1-.1z" class="st8"/><path d="m3.3 3.9 3.3-1.4-1 1.4H3.3z" class="st7"/><path d="m2 6.7 3.7-2.8.3 2.8H2z" class="st3"/><path d="M5.7 3.9 6 6.7l2.8-2.9-3.1.1z" class="st1"/><path d="m2 6.7 1.4-2.8h2.4L2 6.7z" class="st4"/><path d="m11 13.6-1.2 2h3.3l-2.1-2z" class="st7"/><path d="m8.5 13.6 1.4 2 1.2-2H8.5z" class="st0"/><path d="m10.6 10.2.4 3.4 6-1.6-6.4-1.8z" class="st8"/><path d="M13.2 15.6 17 12l-6 1.6 2.2 2z" class="st9"/></svg>
+      </a>
       <h1>
-        <svg fill="none" height="25" viewBox="0 0 171 25" width="171" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
-        <mask id="a" height="8" maskUnits="userSpaceOnUse" width="11" x="15" y="17"><path clip-rule="evenodd" d="m15.0547 17.3882h10.7723v7.207h-10.7723z" fill="#fff" fill-rule="evenodd"/></mask><g clip-rule="evenodd" fill-rule="evenodd"><g fill="var(--color-text)"><path d="m131.861 4.66507-1.374 1.27514h-.012l-1.433-1.27514h-3.072l-1.726 2.0088-1.956-2.0088v13.38343h3.915v-9.20598h1.763l.551.95173v8.25425h3.915l-.404-.461.416.461v-8.25425l.609-.95173h1.21l.51.95173v8.25425h4.032v-10.32834l-3.098-3.05509z"/><path d="m45.2343 12.6127-1.4359 1.3529h-4.1126v-5.18873h4.1126l1.4359 1.57303zm-2.706-7.85309h-2.8425l-1.9915 1.92063-1.998-1.92063v17.84799h3.9895v-4.6244l-.0532-.0555v-.0007l.0532.0562h5.6551l3.4834-3.0551v-6.49921l-3.4834-3.66928z"/><path d="m77.2059 12.9893v-3.49693h4.2249v-3.93985h-4.2249l-2.0167-2.19928.0017-.00168-2.021-1.96707v13.65941l3.9437 3.0046h5.0169v-3.9509h-3.2703z"/><path d="m66.7636 14.1107h-4.1518l-1.4498-1.5417v-2.2185l1.4498-1.32627h4.1518zm0-9.0243h-5.7096l-3.5169 2.99451v6.37069l3.5169 3.5969h5.7096l2.011-1.8826 2.017 1.8826v-9.02427l-2.017-1.89538z"/><path d="m113.625 5.04334-2.016 2.0623-2.02-2.0623v13.00516h4.036v-9.0542h4.161l2.498 2.5627v-4.45674l-2.601-2.09428z"/><path d="m84.0986 3.49927v14.54923h4.0363v-4.835h4.2249v-3.9399h-4.2249v-3.49697l1.6544-1.33067h3.2699v-3.951301h-5.0166z"/><path d="m53.0157 2.6936.0014-.00134-2.021-2.197601v8.777591l.0083.00875v.00135l-.0083-.00909v8.77524h4.0363v-8.77524l-.003.00337v-.00135l.003-.00303v-4.39014l-.001.00135z"/><path d="m103.833 12.5424-1.501 1.5491h-2.7414l-1.5016-1.5491v-2.2286l1.5016-1.33231h2.7414l1.501 1.33231zm-5.8556-7.51792-3.6434 3.00865v6.40167l3.6434 3.6137h5.9676l3.643-3.6137v-6.40167l-3.643-3.00865z"/><path d="m150.889 9.14538v3.25032l-1.425 2.0347h-2.954l-1.425-2.0347v-6.50027l1.425-2.03505h2.954l1.425 2.03505zm-5.919-8.737207-3.736 3.482837v10.50839l3.736 3.4832h6.033l3.736-3.4832v-10.50839l-3.736-3.482837z"/><path d="m167.448 7.52596h-.005l-.127-.09726.029.09726h-.061l.032-.09726h-5.183l-1.054-1.0012v-1.3512l1.054-1.20178h3.326l2.088 1.83413 3.186-1.80317-3.735-3.496972h-6.033l-3.736 3.496972v4.32149l3.417 2.63543h3.421v.0121h1.409l1.407.9891v1.3515l-1.424 1.2015h-2.956l-2.087-1.8338-3.187 1.8028 3.735 3.497h6.034l3.735-3.497v-4.3218l-3.285-2.5338z"/></g><path d="m4.70002.149719h2.43672l-3.83108 3.921011z" fill="#56b146"/><path d="m2.00038 10.5222s-1.9688335-3.64708-2.00038-3.64372l1.95688.01043z" fill="#c3233d"/><path d="m1.95703 6.88889 4.02467-.00068-3.98117 3.63399z" fill="#d5246a"/><path d="m5.82842 13.7166.15275-6.82839-3.98117 3.63399z" fill="#e77a2b"/><path d="m10.6388 10.3186-4.81165 3.398.15441-6.82839z" fill="#fac922"/><path d="m13.6277 5.82074-7.64625 1.0675 4.65725 3.43036z" fill="#56b146"/><path d="m11.8604.149719-3.08989 3.841251 4.85749 1.82976z" fill="#c3233d"/><path d="m19.1124.149719h-7.252l1.7676 5.671011z" fill="#487fb5"/><path d="m5.98145 6.88823 7.64625-1.0675-4.85754-1.82976z" fill="#d5246a"/><path d="m20.4868 5.82073h-6.8589l5.4845-5.671011z" fill="#23477b"/><path d="m17.0469 12.1355-6.4082-1.8169 2.9889-4.49786z" fill="#419845"/><path d="m20.4868 5.82074-3.4396 6.31476-3.4193-6.31476z" fill="#014c3d"/><path d="m22.8936 10.3186-5.8467 1.8169 3.4395-6.31476z" fill="#419845"/><path d="m21.829 14.004-4.7821-1.8685 5.8467-1.8169z" fill="#56b146"/><path d="m19.1123.149719 5.2799 4.404611-3.9055 1.2664z" fill="#487fb5"/><path d="m20.4863 5.82072 3.9055-1.2664-1.4983 5.76428z" fill="#008b47"/><path d="m22.8937 10.3186 5.4031 1.0358-6.4677 2.6496z" fill="#014c3d"/><path d="m28.2968 11.3544.0651 5.1999-6.5328-2.5503z" fill="#419845"/><path d="m21.8291 14.004 6.5328 2.5503-4.9359.8339z" fill="#23477b"/><path d="m28.3616 16.5543-2.5353 5.6629-2.4005-4.829z" fill="#56b146"/><g mask="url(#a)"><path d="m23.4268 17.3882 2.4005 4.829-3.6584-2.0616z" fill="#487fb5"/><path d="m25.8273 22.2172-3.6185 2.3982-.0399-4.4598z" fill="#23477b"/><path d="m22.1685 20.1556.0398 4.4598-1.8167-2.8966z" fill="#487fb5"/><path d="m22.2089 24.6154h-4.831l3.0142-2.8966z" fill="#23477b"/><path d="m20.3921 21.7188-3.0142 2.8966.5323-2.8966z" fill="#c3233d"/><path d="m17.9098 21.7188-.5323 2.8966-2.3228-2.8966z" fill="#d5246a"/></g><path d="m17.9098 21.7188h-2.8551l1.6935-2.0047z" fill="#c3233d"/><path d="m15.0547 21.7188v-3.7914l1.6935 1.7867z" fill="#e77a2b"/><path d="m24.3918 4.55432 3.9048 6.80008-5.403-1.0358z" fill="#56b146"/><path d="m17.9097 18.0256-.3872-1.9411 2.8551 1.3054z" fill="#419845"/><path d="m15.0547 17.9274 2.8551.0982-.3872-1.9411z" fill="#56b146"/><path d="m16.7482 19.7141 1.1616-1.6885-2.8551-.0982z" fill="#fac922"/><path d="m6.6543 2.67443 2.11627 1.31654-.10261-2.60212z" fill="#419845"/><path d="m6.6543 2.67443 2.01366-1.28558-1.5305-1.239131z" fill="#56b146"/><path d="m3.30566 4.07073 3.83108-3.921011-.48316 2.524711z" fill="#008b47"/><path d="m8.77013 3.99098-2.11627-1.31654-.98491 1.37442z" fill="#014c3d"/><path d="m3.30566 4.07074 3.34792-1.3963-.98492 1.37442z" fill="#419845"/><path d="m1.95703 6.88891 3.71219-2.84005.31248 2.83937z" fill="#e77a2b"/><path d="m5.66895 4.04885.31247 2.83938 2.78871-2.89726z" fill="#c3233d"/><path d="m1.95703 6.88891 1.34919-2.81818 2.363-.02187z" fill="#fac922"/><path d="m11.0079 13.7099-1.16122 2.0044h3.34462z" fill="#419845"/><path d="m8.49121 13.7166 1.35517 1.9977 1.16122-2.0044z" fill="#56b146"/><path d="m10.6387 10.3186.3689 3.3913 6.0393-1.5744z" fill="#014c3d"/><path d="m13.1912 15.7143 3.8559-3.5788-6.0393 1.5744z" fill="#008b47"/></g></svg>
-        <span class="stamp">platformOS development tools</span>
+        <a href="/">
+          <svg class="logotype" xmlns="http://www.w3.org/2000/svg" xml:space="preserve" style="enable-background:new 0 0 135 22.2" viewBox="0 0 135 22.2" fill="currentColor"><path d="m96.2 4.3-1.4 1.3-1.4-1.3h-3.1l-1.7 2-2-2v13.4h3.9V8.4h1.8l.6 1v8.3h3.9l-.4-.5.4.5V9.4l.6-1h1.2l.5 1v8.3h4V7.3l-3.1-3h-3.8zM9.5 12.2l-1.4 1.4H4V8.4h4.1L9.5 10v2.2zM6.8 4.4H4L2 6.3 0 4.4v17.8h4v-4.6l-.1-.1.1.1h5.7l3.5-3.1V8L9.6 4.4H6.8zm34.7 8.2V9.1h4.2v-4h-4.2l-2-2.2-2-2v13.7l3.9 3h5v-4h-3.3l-1.6-1zm-10.4 1.1h-4.2l-1.5-1.5V9.9l1.5-1.3h4.2v5.1zm0-9h-5.7l-3.5 3V14l3.5 3.6h5.7l2-1.9 2 1.9v-9l-2-1.9-2-2zm46.8-.1-2 2.1-2-2.1v13h4v-9h4.2l2.5 2.6V6.7L82 4.6h-4.1zM48.4 3.1v14.5h4v-4.8h4.2V8.9h-4.2V5.4L54.1 4h3.3V0h-5l-4 3.1zm-31.1-.8-2-2.2v17.6h4V4.5l-2-2.2zm50.8 9.8-1.5 1.6h-2.7l-1.5-1.6V9.9l1.5-1.3h2.7l1.5 1.3v2.2zm-5.8-7.5-3.6 3V14l3.6 3.6h6l3.6-3.6V7.6l-3.6-3h-6zm52.9 4.1V12l-1.4 2h-3l-1.4-2V5.5l1.4-2h3l1.4 2v3.2zM109.3 0l-3.7 3.5V14l3.7 3.5h6L119 14V3.5L115.3 0h-6zm22.5 7.1-.2-.1v.1h-.1V7h-5.2l-1.1-1V4.7l1.1-1.2h3.3l2.1 1.8 3.2-1.8-3.6-3.5h-6l-3.7 3.5v4.3l3.4 2.6h4.8l1.4 1v1.4l-1.4 1.2h-3l-2.1-1.8-3.2 1.8 3.7 3.5h6L135 14V9.7l-3.2-2.6z"/></svg>
+          <span class="label">platformOS development tools</span>
+        </a>
+        <span class="instance" class:offline={!$state.online}>
+        {#if $state.online === undefined}
+          connecting…
+        {:else if $state.online === false}
+          disconnected
+        {:else}
+          <a href={$state.online?.MPKIT_URL}>
+            {$state.online?.MPKIT_URL.replace('https://', '')}
+          </a>
+        {/if}
       </h1>
-    </a>
+    </div>
 
     <nav>
       <ul>